EKIRJASTO-282 Remove placeholders for other book lists
