+++ Palace/AppInfrastructure/TPPRootTabBarController.h
@@ -13,7 +13,7 @@
 @property (readonly) TPPR2Owner *r2Owner;
 @property(nonatomic, readonly) TPPCatalogNavigationController *catalogNavigationController;
 @property (nonatomic, readonly) UIViewController *loansAndHoldsViewController;
-@property (nonatomic, readonly) UIViewController *favoritesAndReadViewController;
+@property (nonatomic, readonly) UIViewController *favoritesMainViewController;
 @property (nonatomic, readonly) UIViewController *settingsViewController;
 
 - (void)showAndReloadCatalogViewController;
